Started to work on step 3
Improving manual entities generation.
Fixed gate position on level1
Createad abstract class for ghosts
Now the player is able to interact with ghosts
Correctly using the intended method getNextCurrentCells() instead of getNeighbourPosition()
Implmemented a method to make the player lose a life.
Now the bonus contains a variable for its duration in seconds