Tons of refactoring
Removed TODO not created by me in area graph.
Moved all the code for the random genereation of the maze in a separate game HyperPacman.
Now the hud in SuperPacmanPlayer shows you how long you are going to be invincible.
Now every level has it's own spawn coordinate, scale factoro and next level accessible statically.
Refactored the GetRandomPoint in radius that still needs to be fixed.
Created the GhostInteractionVisitor to handle view interactions